+++ packages/server/src/index.ts
@@ -1,5 +1,5 @@
 export { aap } from "./server";
-export type { ServerHandler } from "./server";
+export type { Handler } from "./server";
 export { ModelProvider, AiModelProvider } from "./model";
 export { Session } from "./session";
 export type { TurnMessages } from "./session";